api/requirements.txt
torch
pillow
torchvision
flask
